Update MongoDB connection string and initialization script
The MongoDB connection string in the deployment.yaml file has been altered to directly connect to the database instead of using a secret reference. Also, an initialization script has been added to the values.yaml to create a database, user, roles and populate basic data. These changes enhance the initialization and connection setup process for the MongoDB database.